// polytrader-console/src/lib/api/models/markets-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Polytrader Control API
 * Control plane API for Polytrader platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { MarketInfoResponse } from './market-info-response';

/**
 * Markets list response.  Markets are ordered by latest_tick_ts descending (newest first). Markets with null latest_tick_ts appear last.
 */
export interface MarketsResponse {
    /**
     * List of markets (ordered newest first)
     */
    'markets': Array<MarketInfoResponse>;
    /**
     * Number of markets
     */
    'count': number;
}