// polytrader-console/src/lib/api/models/market-tick-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Polytrader Control API
 * Control plane API for Polytrader platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Market tick response model.  Represents a single market tick with price data and timestamps. Used for both latest tick and historical ticks endpoints.
 */
export interface MarketTickResponse {
    /**
     * Unique tick identifier
     */
    'tick_id': string;
    /**
     * Wall-clock timestamp (UTC)
     */
    'ts_wall': string;
    /**
     * Monotonic timestamp
     */
    'ts_mono': number;
    /**
     * Market identifier
     */
    'market_slug': string;
    /**
     * Market outcome: UP or DOWN
     */
    'outcome': string;
    /**
     * Best bid price (0-1 range)
     */
    'best_bid': string;
    /**
     * Best ask price (0-1 range)
     */
    'best_ask': string;
    /**
     * Mid-market price
     */
    'mid': string;
    /**
     * Bid-ask spread
     */
    'spread': string;
    /**
     * Spread in basis points
     */
    'spread_bps': string;
}