// polytrader-console/src/lib/api/models/performance-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Polytrader Control API
 * Control plane API for Polytrader platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ClosedTradeItem } from './closed-trade-item';
// May contain unused imports in some cases
// @ts-ignore
import type { PerformanceSummary } from './performance-summary';

/**
 * Past performance for a strategy: summary + paginated closed trades.
 */
export interface PerformanceResponse {
    /**
     * Aggregates over returned items
     */
    'summary': PerformanceSummary;
    /**
     * Closed trades (newest first)
     */
    'items'?: Array<ClosedTradeItem>;
    'next_cursor'?: string | null;
}