// polytrader-console/src/lib/api/models/performance-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Polytrader Control API
 * Control plane API for Polytrader platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Aggregate performance metrics for the returned closed trades.  Computed from the items in this response (page-scoped). When items are empty, total_trades=0, total_realized_pnl=0, win_rate_pct=None.
 */
export interface PerformanceSummary {
    /**
     * Sum of P&L over returned trades (USD)
     */
    'total_realized_pnl': number;
    /**
     * Number of trades in this page
     */
    'total_trades': number;
    'win_rate_pct'?: number | null;
    'current_drawdown'?: number | null;
    'max_drawdown'?: number | null;
}