// polytrader-console/src/lib/api/models/strategy-orders-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Polytrader Control API
 * Control plane API for Polytrader platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { StrategyOrderItem } from './strategy-order-item';

/**
 * Paginated list of orders for a strategy.
 */
export interface StrategyOrdersResponse {
    /**
     * Order records (newest first)
     */
    'items': Array<StrategyOrderItem>;
    'next_cursor'?: string | null;
}